backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
SQLAlchemy==2.0.23